Added functionality and completed the "Starred" activity
Fixed errors and fully finished the starred activity. Allowed card (numbers) up to 9999 to be read and saved by getting the first three initial char values of a starred set (if a comma is not present in the second and third).
Cards can be starred and unstarred on the "learn" panel for the FlashcardsGUI file.